+++ src/Plivo/Resource/Campaign/Campaign.cs
@@ -117,6 +117,21 @@ namespace Plivo.Resource.Campaign
     }
 
     [JsonObject(MemberSerialization.OptIn)]
+    public class UpdateCampaign: Resource
+    {   
+        [JsonProperty("api_id")]
+        public new string ApiId {get; set;}
+
+        [JsonProperty("campaign")]
+        public CampaignResponse Campaign {get; set;}
+
+        public override string ToString()
+        {
+            return JsonConvert.SerializeObject(this, Formatting.Indented);
+        }
+    }
+
+    [JsonObject(MemberSerialization.OptIn)]
     public class CreateCampaign: Resource
     {   
         [JsonProperty("api_id")]
@@ -245,6 +260,27 @@ namespace Plivo.Resource.Campaign
         }
     }
     public class DeleteNumber{
+
+    }
+
+    [JsonObject(MemberSerialization.OptIn)]
+    public class DeleteCampaign: Resource
+    {
+        
+        [JsonProperty("api_id")]
+        public new string ApiId {get; set;}
+
+        [JsonProperty("campaign_id")]
+        public string CampaignId { get; set; }
+
+        [JsonProperty("message")]
+        public string Message {get; set;}
+
+        public override string ToString()
+        {
+            return JsonConvert.SerializeObject(this, Formatting.Indented);
+        }
+
 
     }