+++ tests_netcore/Plivo.NetCore.Test/Resources/TestCampaign.cs
@@ -91,6 +91,51 @@ namespace Plivo.NetCore.Test.Resources
             AssertRequest(request);
         }
 
+
+        [Fact]
+        public void TestCampaignUpdate()
+        {
+            var id = "CCMZZOS";
+            var data = new Dictionary<string, object>()
+            {
+                {"reseller_id", ""},
+                {"description", ""},
+                {"sample1", "sample1"},
+                {"sample2", ""},
+                {"message_flow", ""},
+                {"help_message", ""},
+                {"optin_keywords", ""},
+                {"optin_message", ""},
+                {"optout_keywords", ""},
+                {"optout_message", ""},
+                {"help_keywords", ""},
+                
+            };
+
+            var request =
+                new PlivoRequest(
+                    "POST",
+                    "Account/MAXXXXXXXXXXXXXXXXXX/10dlc/Campaign/" + id + "/",
+                    "",
+                    data);
+
+            var response =
+                System.IO.File.ReadAllText(
+                    SOURCE_DIR + @"../Mocks/campaignUpdateResponse.json"
+                );
+            Setup<UpdateCampaign>(
+                200,
+                response
+            );
+            var resp = Api.Campaign.Update(id, sample1:"sample1");
+            // Assert.Empty(
+            //     ComparisonUtilities.Compare(
+            //         response,
+            //         Api.Campaign.Update(campaign_id:id, sample1:"sample1")));
+
+            AssertRequest(request);
+        }
+
         [Fact]
         public void TestCampaignGet()
         {
@@ -167,6 +212,29 @@ namespace Plivo.NetCore.Test.Resources
             );
             var resp = Api.Campaign.GetNumber(campaignID, number);
            
+            AssertRequest(request);
+        }
+
+        [Fact]
+        public void TestCampaignDelete()
+        {
+            var id = "CUU5RCB";
+            var request =
+                new PlivoRequest(
+                    "DELETE",
+                    "Account/MAXXXXXXXXXXXXXXXXXX/10dlc/Campaign/" + id + "/",
+                    "");
+
+            var response =
+                System.IO.File.ReadAllText(
+                    SOURCE_DIR + @"../Mocks/campaignDeleteResponse.json"
+                );
+            Setup<DeleteResponse<DeleteCampaign>>(
+                200,
+                response
+            );
+            
+            var resp = Api.Campaign.Delete(id);
             AssertRequest(request);
         }
     }